jcr: separate file for CompressionContext
